+++ src/components/shared/scroll-indicator.vue
@@ -1,38 +1,10 @@
 <template>
-  <div class="scroll-indicator" :style="{ '--scroll': scroll }"></div>
+  <div class="scroll-indicator" />
 </template>
 
 <script>
-import { ref, onMounted, onUnmounted } from "@vue/composition-api";
-
 export default {
   name: "u-scroll-indicator",
-  setup() {
-    const scroll = ref(0);
-
-    function onScroll() {
-      requestAnimationFrame(() => {
-        scroll.value =
-          window.pageYOffset /
-          (document.body.offsetHeight - window.innerHeight);
-      });
-    }
-
-    onMounted(() => {
-      if (window) {
-        window.addEventListener("scroll", onScroll);
-      }
-    });
-    onUnmounted(() => {
-      if (window) {
-        window.removeEventListener("scroll", onScroll);
-      }
-    });
-
-    return {
-      scroll,
-    };
-  },
 };
 </script>
