+++ src/layouts/default.vue
@@ -1,22 +1,25 @@
 <template>
-  <div id="app">
-    <u-skip-link />
+  <u-scroll-custom-property>
+    <div id="app">
+      <u-skip-link />
 
-    <u-scroll-indicator />
+      <u-scroll-indicator />
 
-    <u-navigation class="navigation" />
+      <u-navigation class="navigation" />
 
-    <transition name="fade" appear>
-      <main id="main" class="main">
-        <slot />
-      </main>
-    </transition>
+      <transition name="fade" appear>
+        <main id="main" class="main">
+          <slot />
+        </main>
+      </transition>
 
-    <u-footer class="footer" />
-  </div>
+      <u-footer class="footer" />
+    </div>
+  </u-scroll-custom-property>
 </template>
 
 <script>
+import scrollCustomProperty from "~/components/shared/scroll-custom-property.vue";
 import scrollIndicator from "~/components/shared/scroll-indicator.vue";
 import skipLink from "~/components/shared/skip-link.vue";
 import footer from "~/components/footer.vue";
@@ -24,6 +27,7 @@ import navigation from "~/components/navigation.vue";
 
 export default {
   components: {
+    "u-scroll-custom-property": scrollCustomProperty,
     "u-scroll-indicator": scrollIndicator,
     "u-skip-link": skipLink,
     "u-footer": footer,
